fix: adjusted vpc to only deploy public subnet with all instances
Ran into too much complexity with private subnets and needed to adjust to have our instances in a public subnet. We understand for future, this will need to be resolved for extra security, however due to time constraints and the scope of our project, we decided to forgoe private subnets.

Co-authored-by: Garrett <gjcochran@users.noreply.github.com>